[probes.starlark] assert.http_status: add optional msg; split builtins.go
- assert.http_status now takes an optional msg= that is appended to the
  failure error, so scripts can log context (e.g. dynamic headers) that
  produced the failed request.
- Split builtins.go into one file per module (http, assert, vars, log,
  state, print_metric). builtins.go keeps the registry plus the shared
  Starlark<->Go conversion helpers.
