+++ probes/starlark/builtins.go
@@ -15,18 +15,14 @@
 package starlark
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"io"
-	"net/http"
-	"strings"
 
 	starlarklib "go.starlark.net/starlark"
-	"go.starlark.net/starlarkstruct"
 )
 
 // builtins returns the predeclared globals available to every script.
+// Each module lives in its own builtins_<name>.go file; this file is the
+// registry plus the Starlark<->Go conversion shared by several of them.
 func builtins(vars map[string]string) starlarklib.StringDict {
 	return starlarklib.StringDict{
 		"http":         httpModule(),
@@ -36,372 +32,6 @@ func builtins(vars map[string]string) starlarklib.StringDict {
 		"state":        stateModule(),
 		"print_metric": starlarklib.NewBuiltin("print_metric", printMetric),
 	}
-}
-
-// ----------------------------------------------------------------------------
-// http module
-
-func httpModule() *starlarkstruct.Module {
-	return &starlarkstruct.Module{
-		Name: "http",
-		Members: starlarklib.StringDict{
-			"get":  starlarklib.NewBuiltin("http.get", httpGet),
-			"post": starlarklib.NewBuiltin("http.post", httpPost),
-		},
-	}
-}
-
-func httpGet(thread *starlarklib.Thread, _ *starlarklib.Builtin, args starlarklib.Tuple, kwargs []starlarklib.Tuple) (starlarklib.Value, error) {
-	var url string
-	var headers *starlarklib.Dict
-	if err := starlarklib.UnpackArgs("http.get", args, kwargs,
-		"url", &url,
-		"headers?", &headers,
-	); err != nil {
-		return nil, err
-	}
-	return doHTTP(thread, "GET", url, headers, nil, nil)
-}
-
-func httpPost(thread *starlarklib.Thread, _ *starlarklib.Builtin, args starlarklib.Tuple, kwargs []starlarklib.Tuple) (starlarklib.Value, error) {
-	var url string
-	var headers *starlarklib.Dict
-	var body starlarklib.Value
-	var jsonArg starlarklib.Value
-	if err := starlarklib.UnpackArgs("http.post", args, kwargs,
-		"url", &url,
-		"headers?", &headers,
-		"body?", &body,
-		"json?", &jsonArg,
-	); err != nil {
-		return nil, err
-	}
-	return doHTTP(thread, "POST", url, headers, body, jsonArg)
-}
-
-func doHTTP(thread *starlarklib.Thread, method, url string, headers *starlarklib.Dict, body, jsonArg starlarklib.Value) (starlarklib.Value, error) {
-	var reqBody io.Reader
-	contentType := ""
-	switch {
-	case jsonArg != nil:
-		raw, err := starlarkToGo(jsonArg)
-		if err != nil {
-			return nil, fmt.Errorf("http.%s: encoding json arg: %v", strings.ToLower(method), err)
-		}
-		buf, err := json.Marshal(raw)
-		if err != nil {
-			return nil, fmt.Errorf("http.%s: encoding json arg: %v", strings.ToLower(method), err)
-		}
-		reqBody = bytes.NewReader(buf)
-		contentType = "application/json"
-	case body != nil:
-		s, ok := starlarklib.AsString(body)
-		if !ok {
-			if b, ok := body.(starlarklib.Bytes); ok {
-				reqBody = bytes.NewReader([]byte(b))
-			} else {
-				return nil, fmt.Errorf("http.%s: body must be string or bytes", strings.ToLower(method))
-			}
-		} else {
-			reqBody = strings.NewReader(s)
-		}
-	}
-
-	req, err := http.NewRequestWithContext(ctxFromThread(thread), method, url, reqBody)
-	if err != nil {
-		return nil, err
-	}
-	if contentType != "" {
-		req.Header.Set("Content-Type", contentType)
-	}
-	if headers != nil {
-		for _, item := range headers.Items() {
-			k, ok1 := starlarklib.AsString(item[0])
-			v, ok2 := starlarklib.AsString(item[1])
-			if !ok1 || !ok2 {
-				return nil, fmt.Errorf("http.%s: headers keys and values must be strings", strings.ToLower(method))
-			}
-			req.Header.Set(k, v)
-		}
-	}
-
-	resp, err := httpClientFromThread(thread).Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	respBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	// Lossy: multi-valued headers (e.g. Set-Cookie) are joined with ", ".
-	// This is wrong per RFC 7230 for Set-Cookie specifically, but it keeps
-	// the Starlark-side type a flat dict[string,string]. Phase 2 should
-	// expose the raw http.Header (multi-map) when scripts need it.
-	hdr := starlarklib.NewDict(len(resp.Header))
-	for k, v := range resp.Header {
-		_ = hdr.SetKey(starlarklib.String(k), starlarklib.String(strings.Join(v, ", ")))
-	}
-	hdr.Freeze()
-
-	return &response{
-		status:  resp.StatusCode,
-		headers: hdr,
-		body:    respBody,
-	}, nil
-}
-
-// ----------------------------------------------------------------------------
-// Response value
-
-type response struct {
-	status  int
-	headers *starlarklib.Dict
-	body    []byte
-}
-
-var _ starlarklib.Value = (*response)(nil)
-var _ starlarklib.HasAttrs = (*response)(nil)
-
-func (r *response) String() string {
-	return fmt.Sprintf("<response status=%d size=%d>", r.status, len(r.body))
-}
-func (r *response) Type() string            { return "Response" }
-func (r *response) Freeze()                 {}
-func (r *response) Truth() starlarklib.Bool { return starlarklib.Bool(r.status > 0) }
-func (r *response) Hash() (uint32, error)   { return 0, fmt.Errorf("Response is unhashable") }
-
-func (r *response) Attr(name string) (starlarklib.Value, error) {
-	switch name {
-	case "status":
-		return starlarklib.MakeInt(r.status), nil
-	case "headers":
-		return r.headers, nil
-	case "body":
-		return starlarklib.Bytes(r.body), nil
-	case "json":
-		return starlarklib.NewBuiltin("Response.json", r.jsonMethod).BindReceiver(r), nil
-	}
-	return nil, nil
-}
-
-func (r *response) AttrNames() []string {
-	return []string{"status", "headers", "body", "json"}
-}
-
-func (r *response) jsonMethod(_ *starlarklib.Thread, _ *starlarklib.Builtin, args starlarklib.Tuple, kwargs []starlarklib.Tuple) (starlarklib.Value, error) {
-	if len(args) != 0 || len(kwargs) != 0 {
-		return nil, fmt.Errorf("Response.json: takes no arguments")
-	}
-	var v interface{}
-	if err := json.Unmarshal(r.body, &v); err != nil {
-		return nil, fmt.Errorf("Response.json: %v", err)
-	}
-	return goToStarlark(v)
-}
-
-// ----------------------------------------------------------------------------
-// vars module
-
-func varsModule(vars map[string]string) *starlarkstruct.Module {
-	get := func(_ *starlarklib.Thread, _ *starlarklib.Builtin, args starlarklib.Tuple, kwargs []starlarklib.Tuple) (starlarklib.Value, error) {
-		var name string
-		var dflt starlarklib.Value = starlarklib.None
-		if err := starlarklib.UnpackArgs("vars.get", args, kwargs,
-			"name", &name,
-			"default?", &dflt,
-		); err != nil {
-			return nil, err
-		}
-		if v, ok := vars[name]; ok {
-			return starlarklib.String(v), nil
-		}
-		return dflt, nil
-	}
-	return &starlarkstruct.Module{
-		Name: "vars",
-		Members: starlarklib.StringDict{
-			"get": starlarklib.NewBuiltin("vars.get", get),
-		},
-	}
-}
-
-// ----------------------------------------------------------------------------
-// assert module
-
-func assertModule() *starlarkstruct.Module {
-	return &starlarkstruct.Module{
-		Name: "assert",
-		Members: starlarklib.StringDict{
-			"http_status": starlarklib.NewBuiltin("assert.http_status", assertHTTPStatus),
-		},
-	}
-}
-
-func assertHTTPStatus(_ *starlarklib.Thread, _ *starlarklib.Builtin, args starlarklib.Tuple, kwargs []starlarklib.Tuple) (starlarklib.Value, error) {
-	var resp starlarklib.Value
-	var expected int
-	if err := starlarklib.UnpackArgs("assert.http_status", args, kwargs,
-		"response", &resp,
-		"expected", &expected,
-	); err != nil {
-		return nil, err
-	}
-	r, ok := resp.(*response)
-	if !ok {
-		return nil, fmt.Errorf("assert.http_status: first argument must be a Response, got %s", resp.Type())
-	}
-	if r.status != expected {
-		return nil, fmt.Errorf("assert.http_status: expected %d, got %d", expected, r.status)
-	}
-	return starlarklib.None, nil
-}
-
-// ----------------------------------------------------------------------------
-// log module
-//
-// log.{info,warn,error,debug}(msg) routes through the per-target logger
-// stashed on the thread by runProbe (or the probe-level logger during
-// load-time evaluation in newRuntime). Single-string signature; scripts
-// build composite messages with Starlark's % operator before calling.
-
-func logModule() *starlarkstruct.Module {
-	return &starlarkstruct.Module{
-		Name: "log",
-		Members: starlarklib.StringDict{
-			"info":  starlarklib.NewBuiltin("log.info", logAt("info")),
-			"warn":  starlarklib.NewBuiltin("log.warn", logAt("warn")),
-			"error": starlarklib.NewBuiltin("log.error", logAt("error")),
-			"debug": starlarklib.NewBuiltin("log.debug", logAt("debug")),
-		},
-	}
-}
-
-func logAt(level string) func(*starlarklib.Thread, *starlarklib.Builtin, starlarklib.Tuple, []starlarklib.Tuple) (starlarklib.Value, error) {
-	return func(thread *starlarklib.Thread, _ *starlarklib.Builtin, args starlarklib.Tuple, kwargs []starlarklib.Tuple) (starlarklib.Value, error) {
-		var msg string
-		if err := starlarklib.UnpackArgs("log."+level, args, kwargs, "msg", &msg); err != nil {
-			return nil, err
-		}
-		l := loggerFromThread(thread)
-		switch level {
-		case "info":
-			l.Info(msg)
-		case "warn":
-			l.Warning(msg)
-		case "error":
-			l.Error(msg)
-		case "debug":
-			l.Debug(msg)
-		}
-		return starlarklib.None, nil
-	}
-}
-
-// ----------------------------------------------------------------------------
-// state module
-//
-// The bucket lives on sched.RunProbeForTargetRequest.TargetState — created on
-// the first run for a target, freed automatically when the target disappears
-// from discovery (scheduler cancels the goroutine and drops the runReq).
-// Single-goroutine-per-target is a scheduler invariant, so no internal lock.
-//
-// Values round-trip through starlarkToGo / goToStarlark, giving copy-on-get
-// for free. Tuples preserve their identity (not flattened to lists) via the
-// starlarkTuple named type. stateMaxKeys caps unique-key growth (key count,
-// not byte size — a script that grows a single value unboundedly is not
-// protected against). Big-ints that overflow int64 round-trip lossily
-// through float64 (see the int64 case in goToStarlark below).
-
-const stateMaxKeys = 1024
-
-type stateBucket struct {
-	values map[string]interface{}
-}
-
-func newStateBucket() *stateBucket {
-	return &stateBucket{values: make(map[string]interface{})}
-}
-
-func stateModule() *starlarkstruct.Module {
-	return &starlarkstruct.Module{
-		Name: "state",
-		Members: starlarklib.StringDict{
-			"get": starlarklib.NewBuiltin("state.get", stateGet),
-			"set": starlarklib.NewBuiltin("state.set", stateSet),
-		},
-	}
-}
-
-func stateGet(thread *starlarklib.Thread, _ *starlarklib.Builtin, args starlarklib.Tuple, kwargs []starlarklib.Tuple) (starlarklib.Value, error) {
-	var key string
-	var dflt starlarklib.Value = starlarklib.None
-	if err := starlarklib.UnpackArgs("state.get", args, kwargs,
-		"key", &key,
-		"default?", &dflt,
-	); err != nil {
-		return nil, err
-	}
-	bucket := stateBucketFromThread(thread)
-	v, ok := bucket.values[key]
-	if !ok {
-		return dflt, nil
-	}
-	return goToStarlark(v)
-}
-
-func stateSet(thread *starlarklib.Thread, _ *starlarklib.Builtin, args starlarklib.Tuple, kwargs []starlarklib.Tuple) (starlarklib.Value, error) {
-	var key string
-	var value starlarklib.Value
-	if err := starlarklib.UnpackArgs("state.set", args, kwargs,
-		"key", &key,
-		"value", &value,
-	); err != nil {
-		return nil, err
-	}
-	gv, err := starlarkToGo(value)
-	if err != nil {
-		return nil, fmt.Errorf("state.set: %v", err)
-	}
-	bucket := stateBucketFromThread(thread)
-	if _, exists := bucket.values[key]; !exists && len(bucket.values) >= stateMaxKeys {
-		return nil, fmt.Errorf("state.set: bucket exceeds max keys (%d)", stateMaxKeys)
-	}
-	bucket.values[key] = gv
-	return starlarklib.None, nil
-}
-
-// ----------------------------------------------------------------------------
-// print_metric builtin
-//
-// print_metric(line) hands the line to a metrics/payload Parser via a
-// per-run callback installed on the thread by runProbe. Streaming, not
-// buffered: each call dispatches its EventMetrics immediately, matching
-// external probe behavior. Metrics emitted before a script error survive.
-//
-// Line format is exactly what the parser already accepts (cloudprober's
-// payload syntax), so distributions, GAUGE/CUMULATIVE kind, in-cloudprober
-// aggregation, label syntax, and JSON / header metrics all carry over with
-// zero re-projection in Starlark:
-//
-//	print_metric("items_in_cart 5")
-//	print_metric('items_in_cart{user="alice"} 5')
-//	print_metric("checkout_latency_ms 234.5")  # routed to dist if configured
-//	print_metric("dist:sum:899|count:221|lb:-Inf,0.5,2|bc:34,54,121")
-
-// metricEmitFn dispatches one payload-format line. runProbe builds a closure
-// that parses the line and appends the resulting EventMetrics to
-// result.payloadMetrics; module-load installs a no-op.
-type metricEmitFn func(line string)
-
-func printMetric(thread *starlarklib.Thread, _ *starlarklib.Builtin, args starlarklib.Tuple, kwargs []starlarklib.Tuple) (starlarklib.Value, error) {
-	var line string
-	if err := starlarklib.UnpackArgs("print_metric", args, kwargs, "line", &line); err != nil {
-		return nil, err
-	}
-	metricEmitFromThread(thread)(line)
-	return starlarklib.None, nil
 }
 
 // ----------------------------------------------------------------------------
