+++ probes/starlark/starlark_test.go
@@ -486,6 +486,29 @@ def probe(target):
 	assert.Contains(t, results[0].Error.Error(), "must be a Response")
 }
 
+func TestAssert_HTTPStatusFailureMsg(t *testing.T) {
+	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(http.StatusInternalServerError)
+	}))
+	defer srv.Close()
+
+	source := fmt.Sprintf(`
+def probe(target):
+    r = http.get(url = "%s", headers = {"X-Trace": "abc123"})
+    assert.http_status(r, 200, msg = "trace=%%s" %% "abc123")
+`, srv.URL)
+	opts := newOpts(t, hostFromServer(t, srv), source)
+	p := &Probe{}
+	if err := p.Init("script-assert-msg", opts); err != nil {
+		t.Fatalf("Init: %v", err)
+	}
+
+	results := p.RunOnce(context.Background())
+	assert.False(t, results[0].Success)
+	assert.Contains(t, results[0].Error.Error(), "expected 200, got 500")
+	assert.Contains(t, results[0].Error.Error(), "trace=abc123")
+}
+
 // ---------------------------------------------------------------------------
 // Init / config validation
